requirements.txt
Django==2.1
django-allauth==0.38.0
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
